pyproject.toml
[project]
name = "llm-ops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.2.0",
    "langchain-openai>=0.2.0",
    "python-dotenv>=1.0.0",
    "tavily>=1.1.0",
]
